+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add Moonriver Network. Available under `MOVR` or `moonriver` ids.
+- Add option to change user-agent. `--user` or `-u` CLI option. Default user agent is `polkadot-rewards/version`
+
+### Fixed
+- Fix subscan URL. Used to be `network.subscan.io`, new url is `network.api.subscan.io`
+- Fix hitting API Request limit. Change subscan requests to 3 per second.
 
 ## [0.1.0] - 2021-08-08
 ### Added